+++ src/bidiMapper/domains/log/logHelper.spec.ts
