+++ src/bidiMapper/domains/log/logHelper.ts
