+++ src/bidiMapper/domains/script/scriptEvaluator.ts
